third_party/blink/web_tests/platform/generic/fast/events/touch/gesture/gesture-tap-paragraph-end-expected.txt

Tap to the
right:

Verifies that tapping in the whitespace at the end of a paragraph sends mouse events properly to the paragraph node (even though there are no text line boxes here). http://crbug.com/398205

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS point.x is > span.getBoundingClientRect().right
PASS document.elementFromPoint(point.x, point.y).id is "target"
Sending GestureTapDown
Sending GestureShowPress
Sending GestureTap
Received mousemove on target
PASS event.target.id is "target"
PASS document.elementFromPoint(event.clientX, event.clientY).id is "target"
Received mousedown on target
PASS event.target.id is "target"
PASS document.elementFromPoint(event.clientX, event.clientY).id is "target"
Received mouseup on target
PASS event.target.id is "target"
PASS document.elementFromPoint(event.clientX, event.clientY).id is "target"
Received click on target
PASS event.target.id is "target"
PASS document.elementFromPoint(event.clientX, event.clientY).id is "target"
PASS successfullyParsed is true

TEST COMPLETE
